Cambio de Frecuencias de entero a decimal. Se valida que tenga maximo 6 cifras y 3 decimales. Se agrega texto de ejemplo.

+++ censo/models/alarm_central_cuerpo_data.py
@@ -74,10 +74,11 @@ class CuerpoAlarmCentralData(models.Model):
 
 
     #Frecuencias
-    frequency_one=models.IntegerField(null=True, blank=True, verbose_name='Frecuencia 1 (Mhz)')
-    frequency_two=models.IntegerField(null=True, blank=True, verbose_name='Frecuencia 2 (Mhz)')
-    frequency_three=models.IntegerField(null=True, blank=True, verbose_name='Frecuencia 3 (Mhz)')
-    frequency_four=models.IntegerField(null=True, blank=True, verbose_name='Frecuencia 4 (Mhz)')
+    #
+    frequency_one=models.DecimalField(max_digits=6, decimal_places=3,null=True, blank=True, verbose_name='Frecuencia 1 (Mhz)')
+    frequency_two=models.DecimalField(max_digits=6, decimal_places=3,null=True, blank=True, verbose_name='Frecuencia 2 (Mhz)')
+    frequency_three=models.DecimalField(max_digits=6, decimal_places=3,null=True, blank=True, verbose_name='Frecuencia 3 (Mhz)')
+    frequency_four=models.DecimalField(max_digits=6, decimal_places=3,null=True, blank=True, verbose_name='Frecuencia 4 (Mhz)')
 
     #Tipo?
     fk_normalized_frequency = models.ForeignKey('NormalizedFrequency', verbose_name='normalizada', null=True, blank=True)
